+++ projects/ngx-location-picker/src/lib/types/delegate-search.model.ts
@@ -1,4 +1,4 @@
-import { AddressModel } from '../../lib/types/address.model';
+import { AddressModel, LatLngModel } from '../../lib/types/address.model';
 import { CoordinateModel } from '../../lib/types/coordinate.model';
 import { LocationModel } from '../../lib/types/location.model';
 import { CascadingCoordinateRulesModel } from './cascading-rules.model';
@@ -19,4 +19,5 @@ export interface DelegateSearchModel {
   onlyAntwerp: boolean;
   countryCodes: string[];
   bufferSearch?: number;
+  coordinateSearch?: LatLngModel;
 }